src/requirements.txt
asn1crypto==1.5.1
coloredlogs==15.0.1
humanfriendly==10.0
mrtparse==2.1.0
psutil==5.9.0
py-radix==0.10.0
